requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
python-dotenv
faiss-cpu
sentence-transformers
pypdf
pydantic
tenacity